requirements.txt
appdirs==1.4.2
coreapi==2.2.4
coreschema==0.0.4
Django==1.10.5
django-appconf==1.0.2
django-compressor==2.1.1
djangorestframework==3.5.3
gunicorn==19.6.0
itypes==1.1.0
Jinja2==2.9.5
MarkupSafe==0.23
olefile==0.44
openapi-codec==1.3.1
packaging==16.8
Pillow==4.0.0
psycopg2==2.7
Pygments==2.2.0
pyparsing==2.1.10
PyYAML==3.12
rcssmin==1.0.6
requests==2.13.0
rjsmin==1.0.12
simplejson==3.10.0
six==1.10.0
uritemplate==3.0.0
